Add files via upload
This QR Code Generator project is a simple but effective tool that demonstrates the power of JavaScript in generating dynamic content. It allows users to create QR codes easily for any text or URL they wish to share. Whether used for business, personal, or educational purposes, QR codes are an efficient and convenient way to share information quickly and securely.